requirements.txt
fastapi
uvicorn[standard]
langchain-core
langchain-openai
langchain-community
openai
pydantic-settings
python-dotenv
pymongo